Fix trigger expression for size
With these small changes, it is possible to modify the threshold for warning and critial size on a per host basis. Without these, a modified size per host is only reflected in the event description but not in the expression.

And to not forget: Thank you very much for this template! Thoughtfully designed from my perspective.